When connecting to IMAP-Server check for capability CRAM-MD5 first
The dovecot imap server answers with LOGINDISABLED when CAPABILITY is
issued over an unencrypted connection and disable_plaintext_auth=yes
(the default) is in effect. This causes the plugin to abort the
connection even though CRAM-MD5 is supported by the server.

Checking for CRAM-MD5 first lets the connection succeed while still
preventing transmission of unencrypted credentials over an insecure
connection.